AI Manager
This is the artificial intelligence code for a managerial shooting game. The purpose of this class is to use the AI code to initially select a fort to go behind it and then to find its enemy and start shooting, if the fort is destroyed, seek for another empty fort in the nearest and resume shooting again.
this AI just Control player's movement to find fort.